Rename moai module
