+++ lib/gen/roomgen.lua
@@ -32,7 +32,7 @@ local function init_onlevel()
 	unlockslib.init()
 	cooplib.init()
 	liquidlib.init()
-	ankhmoailib.init()
+	moailib.init()
 	doorslib.init()
 	tombstonelib.init()
 	roomdeflib.init()
