+++ lib/gen/tiledef.lua
@@ -929,7 +929,7 @@ module.HD_TILENAME = {
 							end
 						end
 						doorslib.create_door_exit_moai(x+1, y-3, l)
-						ankhmoailib.create_moai_veil(x, y, l)
+						moailib.create_moai_veil(x, y, l)
 					end,
 				},
 			},
