+++ main.lua
@@ -33,7 +33,7 @@ boulderlib = require 'lib.entities.boulder'
 idollib = require 'lib.entities.idol'
 liquidlib = require 'lib.entities.liquid'
 treelib = require 'lib.entities.tree'
-ankhmoailib = require 'lib.entities.ankhmoai'
+moailib = require 'lib.entities.moai'
 doorslib = require 'lib.entities.doors'
 tombstonelib = require 'lib.entities.tombstone'
 flagslib = require 'lib.flags'
